refactor: abstract set_params, get_params, and reset
